test(oracle): run the migration round trip against Oracle by default
generate_migration_record_test.go sat behind //go:build oracle_record, so
its live round trip -- generate the rollback, execute it against a real
Oracle, require the schema returns to its starting state across 34 cases --
never ran in CI. The tag was not guarding container availability: the
package already runs untagged tests on the same SharedOracleContainer. It
was guarding a side effect, the os.WriteFile that refreshes
testdata/generate_migration/*.json.

Drop the tag and gate the recording behind the `record` const the package
already uses for its .sql goldens. The test is renamed
TestGenerateMigrationAgainstOracle after what it asserts rather than the
side effect, and merges into generate_migration_test.go, which removes the
last test file with no source-file pair.

Re-recording is idempotent (stats zeroed, Oracle-allocated identifiers
canonicalized), so enabling `record` produces no fixture churn.

Cost: backend/plugin/schema/oracle goes 96.9s -> 365.6s, against the
30-minute budget shared by the whole backend suite. Split out as its own
commit so it can be dropped or reverted independently of the test
consolidation.
